Move `MongoConfigurationContributor` to the new `spi` subpackage
